Revert "Merge branch 'items#purchase_by_payjp' of https://github.com/aki2626/freemarcket_sample_68C into items#purchase_by_payjp"
This reverts commit 58fae837f27b00844af916aa3bc7464e2550eddf, reversing
changes made to d42f7ecf8a80ea4aa9a4cb24d43de2a9e0ccb1a4.

+++ db/migrate/20200504113916_create_addresses.rb
@@ -7,7 +7,7 @@ class CreateAddresses < ActiveRecord::Migration[5.2]
       t.text :city_number, null: false
       t.text :building_name
       t.references :user
-      t.bigint :phone_number, unique: true
+      t.integer :phone_number, unique: true
       t.timestamps
     end
   end
